+++ app/helpers/shiny_discussion_helper.rb
@@ -19,4 +19,8 @@ module ShinyDiscussionHelper
   def allow_authenticated_comments
     Setting.get( :allowed_to_comment ).downcase == 'authenticated' || allow_pseudonymous_comments
   end
+
+  def recent_comments_by_user( user, count = 10 )
+    Comment.readonly.where( user: user ).recent.limit( count )
+  end
 end
