+++ app/views/shinycms/discussions/profile/_content.html.erb
@@ -1,0 +1,16 @@
+<% comments = recent_comments_by_user( @user_profile, 5 ) %>
+
+<% if comments.present? %>
+<header style="margin: 0;">
+  <h3>
+    <%= t( '.title' ) %>
+  </h3>
+</header>
+<ul style="list-style: none; margin-left: 0;">
+  <% comments.each do |comment| %>
+  <li>
+    <%= link_to ( comment.title || 'Untitled' ), comment.anchored_path %>
+  </li>
+  <% end %>
+</ul>
+<% end %>
