+++ plugins/ShinyNews/app/views/shiny_news/profile/_content.html.erb
@@ -1,0 +1,17 @@
+<% news_posts = recent_news_posts_by_user( @user_profile, 5 ) %>
+
+<% if news_posts.present? %>
+<header style="margin: 0;">
+  <h3>
+    <%#= link_to t( '.title' ), shiny_news.view_news_posts_by_user_path( @user_profile ) %>
+    <%= link_to t( '.title' ), shiny_news.view_news_path %>
+  </h3>
+</header>
+<ul style="list-style: none; margin-left: 0;">
+  <% news_posts.each do |post| %>
+  <li>
+    <%= link_to post.title, post.path %>
+  </li>
+  <% end %>
+</ul>
+<% end %>
