+++ plugins/ShinyProfiles/app/views/shiny_profiles/profiles/show.html.erb
@@ -2,7 +2,7 @@
 
 <%= render partial: 'public' %>
 
-<%#= render partial: 'content' %>
+<%= render partial: 'content' %>
 
 <%= render partial: 'private' %>
